+++ tests/eval_files/trio103.py
@@ -10,31 +10,58 @@ def foo() -> Any:
 
 
 try:
-    pass
+    ...
 except (SyntaxError, ValueError, BaseException):
     raise
-except (SyntaxError, ValueError, trio.Cancelled) as p:  # error: 33, "trio.Cancelled"
-    pass
+
+try:
+    ...
+except (
+    SyntaxError,
+    ValueError,
+    trio.Cancelled,  # error: 4, "trio.Cancelled", ""
+) as p:
+    ...
+
+try:
+    ...
 except (SyntaxError, ValueError):
     raise
+
+try:
+    ...
 except trio.Cancelled as e:
     raise e
+
+try:
+    ...
 except trio.Cancelled as e:
     raise  # acceptable - see https://peps.python.org/pep-0678/#example-usage
-except trio.Cancelled:  # error: 7, "trio.Cancelled"
-    pass
+
+try:
+    ...
+except trio.Cancelled:  # error: 7, "trio.Cancelled", ""
+    ...
 
 # if
-except BaseException as e:  # error: 7, "BaseException"
+try:
+    ...
+except BaseException as e:  # error: 7, "BaseException", " Consider adding an `except trio.Cancelled: raise` before this exception handler."
     if True:
         raise e
     elif True:
-        pass
+        ...
     else:
         raise e
-except BaseException:  # error: 7, "BaseException"
+
+try:
+    ...
+except BaseException:  # error: 7, "BaseException", " Consider adding an `except trio.Cancelled: raise` before this exception handler."
     if True:
         raise
+
+try:
+    ...
 except BaseException:  # safe
     if True:
         raise
@@ -45,36 +72,52 @@ except BaseException:  # safe
 
 # loops
 # raises inside the body are never guaranteed to run and are ignored
-except trio.Cancelled:  # error: 7, "trio.Cancelled"
+try:
+    ...
+except trio.Cancelled:  # error: 7, "trio.Cancelled", ""
     while foo():
         raise
 
 # raise inside else are guaranteed to run, unless there's a break
+try:
+    ...
 except trio.Cancelled:
     while ...:
         ...
     else:
         raise
+
+try:
+    ...
 except trio.Cancelled:
     for _ in "":
         ...
     else:
         raise
-except BaseException:  # error: 7, "BaseException"
+
+try:
+    ...
+except BaseException:  # error: 7, "BaseException", " Consider adding an `except trio.Cancelled: raise` before this exception handler."
     while ...:
         if ...:
             break
         raise
     else:
         raise
-except BaseException:  # error: 7, "BaseException"
+
+try:
+    ...
+except BaseException:  # error: 7, "BaseException", " Consider adding an `except trio.Cancelled: raise` before this exception handler."
     for _ in "":
         if ...:
             break
         raise
     else:
         raise
+
 # ensure we don't ignore previous guaranteed raise (although that's unreachable code)
+try:
+    ...
 except BaseException:
     raise
     for _ in "":
@@ -87,84 +130,169 @@ except BaseException:
 # nested try
 # in theory safe if the try, and all excepts raises - and there's a bare except.
 # But is a very weird pattern that we don't handle.
-except BaseException as e:  # error: 7, "BaseException"
+try:
+    ...
+except BaseException as e:  # error: 7, "BaseException", " Consider adding an `except trio.Cancelled: raise` before this exception handler."
     try:
         raise e
     except ValueError:
         raise e
     except:
         raise e  # TRIO104: 8
+
+try:
+    ...
 except BaseException:  # safe
     try:
-        pass
+        ...
     finally:
         raise
+
 # check that nested non-critical exceptions are ignored
+try:
+    ...
 except BaseException:
     try:
-        pass
+        ...
     except ValueError:
-        pass  # safe
+        ...  # safe
     raise
+
 # check that name isn't lost
+try:
+    ...
 except trio.Cancelled as e:
     try:
-        pass
+        ...
     except BaseException as f:
         raise f
     raise e
+
 # don't bypass raise by raising from nested except
+try:
+    ...
 except trio.Cancelled as e:
     try:
-        pass
+        ...
     except ValueError as g:
         raise g  # TRIO104: 8
     except BaseException as h:
         raise h  # error? currently treated as safe
     raise e
+
 # bare except, equivalent to `except baseException`
-except:  # error: 0, "bare except"
-    pass
 try:
-    pass
+    ...
+except:  # error: 0, "bare except", " Consider adding an `except trio.Cancelled: raise` before this exception handler."
+    ...
+
+try:
+    ...
 except:
     raise
 
 # point to correct exception in multi-line handlers
 my_super_mega_long_exception_so_it_gets_split = SyntaxError
 try:
-    pass
+    ...
 except (
     my_super_mega_long_exception_so_it_gets_split,
     SyntaxError,
-    BaseException,  # error: 4, "BaseException"
+    BaseException,  # error: 4, "BaseException", " Consider adding an `except trio.Cancelled: raise` before this exception handler."
     ValueError,
     trio.Cancelled,  # no complaint on this line
 ):
-    pass
+    ...
 
 # loop over non-empty static collection
+try:
+    ...
 except BaseException as e:
     for i in [1, 2, 3]:
         raise
-except BaseException:  # error: 7, "BaseException"
+
+try:
+    ...
+except BaseException:  # error: 7, "BaseException", " Consider adding an `except trio.Cancelled: raise` before this exception handler."
     for i in [1, 2, 3]:
         ...
-except BaseException:  # error: 7, "BaseException"
+
+try:
+    ...
+except BaseException:  # error: 7, "BaseException", " Consider adding an `except trio.Cancelled: raise` before this exception handler."
     for i in [1, 2, 3]:
         if ...:
             continue
         raise
+
+try:
+    ...
 except BaseException:
     while True:
         raise
-except BaseException:  # error: 7, "BaseException"
+
+try:
+    ...
+except BaseException:  # error: 7, "BaseException", " Consider adding an `except trio.Cancelled: raise` before this exception handler."
     while True:
         if ...:
             break
         raise
+
+try:
+    ...
 except BaseException:
     while True:
         if ...:
             continue
         raise
+
+
+# Issue #106, false alarm on excepts after `Cancelled` has already been handled
+try:
+    ...
+except trio.Cancelled:
+    raise
+except BaseException:  # now silent
+    ...
+except:  # now silent
+    ...
+
+try:
+    ...
+except trio.Cancelled:
+    raise
+except:  # now silent
+    ...
+
+
+try:
+    ...
+except BaseException:
+    raise
+except:  # now silent
+    ...
+
+# don't throw multiple 103's even if `Cancelled` wasn't properly handled.
+try:
+    ...
+except trio.Cancelled:  # error: 7, "trio.Cancelled", ""
+    ...
+except BaseException:  # now silent
+    ...
+except:  # now silent
+    ...
+
+# Check state management of whether cancelled has been handled across nested try's
+try:
+    try:
+        ...
+    except trio.Cancelled:
+        raise
+except trio.Cancelled:  # error: 7, "trio.Cancelled", ""
+    ...
+except:
+    try:
+        ...
+    except trio.Cancelled:  # error: 11, "trio.Cancelled", ""
+        ...